fixup! KTOR-6668 Move internal types from test-host module; fix gradle references for external import
