+++ ktor-server/ktor-server-cio/jvm/test/io/ktor/tests/server/cio/CIOWebSocketTestJvm.kt
@@ -6,6 +6,7 @@ package io.ktor.tests.server.cio
 
 import io.ktor.client.*
 import io.ktor.client.plugins.websocket.*
+import io.ktor.server.base.*
 import io.ktor.server.cio.*
 import io.ktor.server.testing.*
 import kotlinx.coroutines.*
