+++ ktor-server/ktor-server-cio/jvm/test/io/ktor/tests/server/cio/ClassCastExceptionTest.kt
@@ -8,6 +8,7 @@ import io.ktor.client.*
 import io.ktor.client.call.*
 import io.ktor.client.request.*
 import io.ktor.server.application.*
+import io.ktor.server.base.*
 import io.ktor.server.cio.*
 import io.ktor.server.engine.*
 import io.ktor.server.routing.*
