+++ ktor-server/ktor-server-plugins/ktor-server-websockets/jvmAndNix/test/io/ktor/tests/websocket/DefaultWebSocketTest.kt
@@ -4,6 +4,7 @@
 
 package io.ktor.tests.websocket
 
+import io.ktor.server.base.*
 import io.ktor.server.testing.*
 import io.ktor.utils.io.*
 import io.ktor.utils.io.charsets.*
