+++ ktor-server/ktor-server-test-base/build.gradle.kts
@@ -6,7 +6,6 @@ kotlin.sourceSets {
     jvmAndNixMain {
         dependencies {
             api(project(":ktor-server:ktor-server-test-host"))
-            api(project(":ktor-shared:ktor-junit"))
         }
     }
