+++ ktor-server/ktor-server-test-base/jvm/src/io/ktor/server/base/BaseTestJvm.kt
@@ -3,7 +3,7 @@
  * Copyright 2014-2022 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
  */
 
-package io.ktor.server.testing
+package io.ktor.server.base
 
 import io.ktor.junit.*
 import kotlinx.coroutines.*
