+++ ktor-server/ktor-server-test-base/jvmAndNix/src/io/ktor/server/base/BaseTest.kt
@@ -2,7 +2,7 @@
  * Copyright 2014-2022 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
  */
 
-package io.ktor.server.testing
+package io.ktor.server.base
 
 import kotlinx.coroutines.*
 import kotlin.time.*
