+++ ktor-server/ktor-server-test-base/jvmAndNix/src/io/ktor/server/base/EngineTestBase.kt
@@ -1,8 +1,8 @@
 /*
- * Copyright 2014-2021 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
+ * Copyright 2014-2024 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
  */
 
-package io.ktor.server.testing
+package io.ktor.server.base
 
 import io.ktor.client.request.*
 import io.ktor.client.statement.*
