+++ ktor-server/ktor-server-test-host/jvmAndNix/src/io/ktor/server/testing/TestApplicationCall.kt
@@ -12,7 +12,7 @@ import kotlin.coroutines.*
 /**
  * A test application call that is used in [withTestApplication] and [handleRequest].
  */
-class TestApplicationCall(
+public class TestApplicationCall(
     application: Application,
     readResponse: Boolean = false,
     closeRequest: Boolean = true,
