+++ ktor-server/ktor-server-test-suites/jvm/src/io/ktor/server/testing/suites/ServerPluginsTestSuite.kt
@@ -5,6 +5,7 @@
 package io.ktor.server.testing.suites
 
 import io.ktor.server.application.*
+import io.ktor.server.base.*
 import io.ktor.server.engine.*
 import io.ktor.server.request.*
 import io.ktor.server.response.*
