+++ ktor-server/ktor-server-test-suites/jvmAndNix/src/io/ktor/server/testing/suites/HttpServerCommonTestSuite.kt
@@ -12,6 +12,7 @@ import io.ktor.http.*
 import io.ktor.http.content.*
 import io.ktor.server.application.*
 import io.ktor.server.application.hooks.*
+import io.ktor.server.base.*
 import io.ktor.server.engine.*
 import io.ktor.server.http.*
 import io.ktor.server.plugins.*
