+++ settings.gradle.kts
@@ -54,6 +54,7 @@ include(":ktor-server:ktor-server-config-yaml")
 include(":ktor-server:ktor-server-tests")
 include(":ktor-server:ktor-server-host-common")
 include(":ktor-server:ktor-server-test-host")
+include(":ktor-server:ktor-server-test-base")
 include(":ktor-server:ktor-server-test-suites")
 include(":ktor-server:ktor-server-jetty")
 include(":ktor-server:ktor-server-jetty:ktor-server-jetty-test-http2")
@@ -171,5 +172,3 @@ include(":ktor-shared:ktor-websockets")
 include(":ktor-shared:ktor-sse")
 include(":ktor-shared:ktor-junit")
 include(":ktor-java-modules-test")
-include("ktor-server:ktor-server-test-base")
-findProject(":ktor-server:ktor-server-test-base")?.name = "ktor-server-test-base"
